Correcciones
Se realizan correcciones sobre los generadores y sobre el grafo.

+++ Cursada/TP4/tp4/source/MatrizSimetrica.java
@@ -114,7 +114,7 @@ public class MatrizSimetrica {
 		return minant;
 	}
 
-	public double getPorcentajeAdyacencia() {
+	public int getPorcentajeAdyacencia() {
 		return this.getCantidadAristas() * 100 / this.getCantidadAristasTotales();
 	}
